Changed some attributes to smaller ones (memory)

+++ include/game/base/Message.hpp
@@ -54,8 +54,8 @@ namespace grid
     {
     public :
 
-        int type;
-        int health;
+        short int type;
+        short int health;
         Entity * sender;
         bool flag;
